move test from base to SQL suite
