+++ core/src/test/scala/org/apache/spark/sql/delta/ConvertToDeltaSuiteBase.scala
@@ -120,28 +120,6 @@ trait ConvertToDeltaSuiteBase extends ConvertToDeltaSuiteBaseCommons
     }
   }
 
-  test("convert with collectStats set to false") {
-    withTempDir { dir =>
-      withSQLConf(DeltaSQLConf.DELTA_COLLECT_STATS.key -> "true") {
-
-        val tempDir = dir.getCanonicalPath
-        writeFiles(tempDir, simpleDF)
-        convertToDelta(s"parquet.`$tempDir`", collectStats = false)
-        val deltaLog = DeltaLog.forTable(spark, tempDir)
-        val history = io.delta.tables.DeltaTable.forPath(tempDir).history()
-        checkAnswer(
-          spark.read.format("delta").load(tempDir),
-          simpleDF
-        )
-        assert(history.count == 1)
-        val statsDf = deltaLog.unsafeVolatileSnapshot.allFiles
-            .select(from_json($"stats", deltaLog.unsafeVolatileSnapshot.statsSchema)
-            .as("stats")).select("stats.*")
-        assert(statsDf.filter($"numRecords".isNotNull).count == 0)
-      }
-    }
-  }
-
   test("negative case: convert a non-delta path falsely claimed as parquet") {
     Seq("orc", "json", "csv").foreach { format =>
       withTempDir { dir =>
